backend/requirements.txt
alembic==1.11.2
annotated-types==0.5.0
anyio==3.7.1
async-exit-stack==1.0.1
async-generator==1.10
asyncpg==0.28.0
beautifulsoup4==4.12.2
blinker==1.6.3
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.6
databases==0.7.0
exceptiongroup==1.1.3
fastapi==0.100.1
Flask==3.0.0
google==3.0.0
google-api-core==2.12.0
google-auth==2.23.4
google-cloud==0.34.0
google-cloud-core==2.3.3
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
Mako==1.2.4
MarkupSafe==2.1.3
MyApplication==0.1.0
Pillow==10.1.0
protobuf==4.24.4
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.1.1
pydantic_core==2.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
s3transfer==0.6.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==1.4.49
starlette==0.27.0
typing_extensions==4.7.1
urllib3==1.26.16
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
Werkzeug==3.0.1
zipp==3.17.0
